refactoring(handlers): unit test modularization
