+++ handlers/api/api.go
@@ -1,8 +1,11 @@
 package api
 
 import (
+	"log"
+
 	"github.com/aws/aws-sdk-go-v2/aws"
 	"github.com/aws/aws-sdk-go-v2/service/iam"
+	"github.com/rajasoun/aws-hub/provider/credential"
 )
 
 const IAMGetUserCountAPI = "GetUserCountAPI"
@@ -26,8 +29,24 @@ func NewAwsAPI(api string) AwsAPI {
 	return awsAPI
 }
 
-func GetConfig(profile string, isMultipleProfile bool) (aws.Config, error) {
-	credentialLoader := CredentialLoader{}
-	cfg, err := credentialLoader.GetConfig(profile, isMultipleProfile)
+func GetConfigFromFileSystem(profile string, isMultipleProfile bool) (aws.Config, error) {
+	var cfg aws.Config
+	var err error
+	credentialLoader := credential.CredentialLoader{}
+	if isMultipleProfile {
+		cfg, err = credentialLoader.LoadDefaultConfigForProfile(profile)
+		handleErr(err, "AWSConfig For multiple Profile ")
+	} else {
+		cfg, err = credentialLoader.LoadDefaultConfig()
+		handleErr(err, "Default AWSConfig")
+	}
 	return cfg, err
+}
+
+func handleErr(err error, msg string) {
+	if err != nil {
+		log.Println(msg+" Load Failed err = %v", err)
+	} else {
+		log.Println(msg + "loaded successfuly")
+	}
 }
