+++ handlers/api/api_test.go
@@ -58,7 +58,7 @@ func TestGetConfig(t *testing.T) {
 	for _, tt := range tests {
 		t.Run(tt.name, func(t *testing.T) {
 			assert := assert.New(t)
-			cfg, err := GetConfig(tt.profile, tt.isMultipleProfile)
+			cfg, err := GetConfigFromFileSystem(tt.profile, tt.isMultipleProfile)
 			if tt.wantErr {
 				assert.Error(err, "GetConfig() = %v ", err)
 			}
